+++ client/packages/common/src/intl/locales/en/common.json
@@ -1123,6 +1123,7 @@
   "label.properties": "Properties",
   "label.provider-name": "Provider name",
   "label.purchase-cost-price": "Purchase cost",
+  "label.purchase-order": "Purchase Order",
   "label.qty-item-name": "Qty/Item Name",
   "label.quantities": "Quantities",
   "label.quantity": "Quantity",
