--- client/packages/system/src/Name/DetailModal/index.ts
@@ -1,1 +1,0 @@
-export * from './DetailModal';
