+++ client/packages/system/src/Name/Details/index.ts
@@ -1,0 +1,1 @@
+export * from './Details';
