+++ client/packages/system/src/Name/SupplierDetailView/index.ts
@@ -1,0 +1,1 @@
+export * from './SupplierDetailView';
